% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{ac_heatmap_ggplot2}
\alias{ac_heatmap_ggplot2}
\title{plotting only one annotation column for heatmap with ggplot2, ac could be discrete / continues values}
\usage{
ac_heatmap_ggplot2(
  ac,
  group.name,
  fill_levels,
  pal.continuous = c("white", "red"),
  pal.discrete = (ggsci::pal_lancet(palette = c("lanonc"), alpha = 0.6))(9),
  legend.title,
  legend.title.size = 18,
  legend.text.size = 16
)
}
\arguments{
\item{ac}{a dataframe ac for annotation the column for heatmap, row names must be consistant with the heatmap columns; colunms can be  one or more, but only one column used for annotaition setting}

\item{group.name}{the colunm name in ac data frame, picked for annotation}

\item{fill_levels}{if the filling is discrete values, can define the factors orders using levels}

\item{pal.continuous}{if the filling is continue values, using continues pal, a two value vector: c('low color', 'high color'), default c("white", "red")}

\item{pal.discrete}{if the filling is discrete values, using discrete pal, if denfined the length of the pal must be consistant with the length of fill factor level,
default: ggsci::pal_lancet(palette = c("lanonc"), alpha = 0.6)(9)}

\item{legend.title}{title of the ac legend bar, if not given, using the column name of the ac data frame}

\item{legend.title.size}{text size of the ac bar legend title, default 18}

\item{legend.text.size}{text size of the ac bar legend lab, default 16}
}
\value{
a ggplot2 object of one annotation column of a heatmap
}
\description{
plotting only one annotation column for heatmap with ggplot2, ac could be discrete / continues values
}
\examples{
ac_heatmap_ggplot2(ac = ac, group.name = 'group_list')
}
