% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplot_continuous}
\alias{barplot_continuous}
\title{Barplot (dodge) with data presented as long, do not analysis and provide mean or sd or counts. Suitable for y data type of continuous!}
\usage{
barplot_continuous(
  data,
  x,
  sub.x,
  value,
  errbar = T,
  lab_text = F,
  labtext.size = 3,
  xlab_level,
  fill_level,
  bar_width = 0.9,
  title = NULL,
  x_lab = NULL,
  y_lab = "",
  base.size = 15,
  legend.position = "top",
  legend.size = 12,
  xlab.angle = 0,
  pal = (ggsci::pal_lancet(palette = c("lanonc"), alpha = 0.6))(9),
  chinese = F,
  plot.theme
)
}
\arguments{
\item{data}{Dataframe with long form (raw data, do not analysis!); y continuous data}

\item{x}{x column name be calulated , data type must be discrete}

\item{sub.x}{sub x column name be calulated , data type must be discrete (if two varibles need to be caculated for mean/sd)}

\item{value}{Y column name,  data type must be continuous for mean/sd}

\item{errbar}{if  adding error bar, default True}

\item{lab_text}{if  adding text of y value, default False}

\item{labtext.size}{text size the the labs add on the bar}

\item{xlab_level}{levels of x lab}

\item{fill_level}{levels of fill (legend)}

\item{bar_width}{bar width default 0.9}

\item{title}{plot title}

\item{x_lab}{plot x lab}

\item{y_lab}{plot y lab}

\item{base.size}{base text size of the theme default 15}

\item{legend.position}{legend position default top}

\item{legend.size}{legend text size default 12}

\item{xlab.angle}{if rotate the x labs defualt 0}

\item{pal}{color for filling  default pal_lancet}

\item{chinese}{if text in Chinese ,default False}

\item{plot.theme}{theme for plot deault classic}
}
\value{
ggplot plot
}
\description{
Barplot (dodge) with data presented as long, do not analysis and provide mean or sd or counts. Suitable for y data type of continuous!
}
\examples{
 barplot_continuous(data = sur_data, x = 'paper_BRCA_Subtype_PAM50', value = 'total_perMB', lab_text = T, errbar = F)
}
