% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplot_discrete}
\alias{barplot_discrete}
\title{Barplot with data presented as long, do not analysis and provide mean or sd or counts. Suitable for xtab counting results presenting!}
\usage{
barplot_discrete(
  data,
  x,
  count,
  bar_location = "dodge",
  lab_text = T,
  xlab_level,
  y_limit_expand,
  bar_width = 0.9,
  title = NULL,
  x_lab = NULL,
  y_lab = "",
  base.size = 18,
  legend.position = "top",
  legend.size = 12,
  xlab.angle = 0,
  pal = (ggsci::pal_lancet(palette = c("lanonc"), alpha = 0.6))(9),
  chinese = F,
  plot.theme
)
}
\arguments{
\item{data}{Dataframe with long form (raw data, do not analysis!); all discrete data}

\item{x}{x column name be counted , data type must be discrete}

\item{count}{y column name to be counted , data type must be discrete}

\item{bar_location}{bar position could be dodge(default), fill, stack}

\item{lab_text}{adding text of y value on the bar, default True}

\item{xlab_level}{levels of x lab}

\item{y_limit_expand}{expand the y limit, give a numer of percent of y low/up limit, such as c(0, 0.05) for expanding 0\% to lower limit and 5\% to upper limit, default NULL}

\item{bar_width}{bar width default 0.9}

\item{title}{plot title}

\item{x_lab}{plot x lab}

\item{y_lab}{plot y lab}

\item{base.size}{base text size of the theme default 18}

\item{legend.position}{legend position default top}

\item{legend.size}{legend text size default 12}

\item{xlab.angle}{if rotate the x labs defualt 0}

\item{pal}{color for filling  default pal_lancet}

\item{chinese}{if text in Chinese ,default False}

\item{plot.theme}{theme for plot deault classic}
}
\value{
ggplot plot
}
\description{
Barplot with data presented as long, do not analysis and provide mean or sd or counts. Suitable for xtab counting results presenting!
}
\examples{
barplot_discrete(data = sur_data, x = 'tmb', count= 'paper_BRCA_Subtype_PAM50', xlab_level = c('low', 'high'), lab_text = T)
}
