% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplot_identity}
\alias{barplot_identity}
\title{Barplot with data presented identity (as rawdata give mean , sd), subtype shown as bar dodge}
\usage{
barplot_identity(
  data,
  x_value,
  y_value,
  fill_group,
  errbar,
  lab_text = F,
  labtext.size = 3,
  xlab_level,
  fill_level,
  y_limit_expand,
  title = NULL,
  x_lab = NULL,
  y_lab = "",
  bar_width = 0.6,
  base.size = 18,
  legend.position = "top",
  legend.size = 12,
  xlab.angle = 0,
  pal = (ggsci::pal_lancet(palette = c("lanonc"), alpha = 0.6))(9),
  chinese = F,
  plot.theme
)
}
\arguments{
\item{data}{data frame anlysised previously with mean, sd}

\item{x_value}{column name of x value in the dataframe}

\item{y_value}{column name of y value in the dataframe, such as mean}

\item{fill_group}{column name of filling for subtype}

\item{errbar}{column name of error bar}

\item{lab_text}{adding text of y value on the bar, default False}

\item{labtext.size}{text size the the labs add on the bar}

\item{xlab_level}{levels of x lab}

\item{fill_level}{levels of fill lab}

\item{y_limit_expand}{expand the y limit, give a numer of percent of y low/up limit, such as c(0, 0.05) for expanding 0\% to lower limit and 5\% to upper limit, default NULL}

\item{title}{plot title}

\item{x_lab}{plot x lab}

\item{y_lab}{plot y lab}

\item{bar_width}{bar width default 0.6}

\item{base.size}{base text size of the theme default 18}

\item{legend.position}{legend position default top}

\item{legend.size}{legend text size default 12}

\item{xlab.angle}{if rotate the x labs defualt 0}

\item{pal}{color for filling  default pal_lancet}

\item{chinese}{if text in Chinese ,default False}

\item{plot.theme}{theme for plot deault classic}
}
\value{
ggplot2 bar plot
}
\description{
Barplot with data presented identity (as rawdata give mean , sd), subtype shown as bar dodge
}
\examples{
barplot_identity(data = dat, x_value = 'paper_BRCA_Subtype_PAM50', y_value = 'tmb_mean', fill_group = 'paper_BRCA_Subtype_PAM50', errbar = 'tmb_sd')
}
