% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplot_jitter_prism}
\alias{barplot_jitter_prism}
\title{Barplot (dodge, adding points and prism theme) with data presented as long, do not analysis and provide mean or sd or counts. Suitable for y data type of continuous!}
\usage{
barplot_jitter_prism(
  data,
  x,
  value,
  fill,
  xlab_level,
  fill_level,
  jitter = T,
  fill.col = T,
  fill.alpha = 1,
  title = NULL,
  x_lab = NULL,
  y_lab = "",
  point.size = 3,
  title.text.size = 20,
  xlab.text.size = 18,
  ylab.text.size = 18,
  xtitle.text.size = 18,
  ytitle.text.size = 18,
  xlab.angle = 0,
  legend.position = "top",
  legend.text.size = 12,
  pal = (ggsci::pal_lancet(palette = c("lanonc")))(9),
  plot.theme
)
}
\arguments{
\item{data}{Dataframe with long form (raw data, do not analysis!); y continuous data}

\item{x}{x column name be counted , data type must be discrete}

\item{value}{Y column name,  data type must be continuous for mean/sd}

\item{fill}{filling column, data type must be discrete, for legend filling column.}

\item{xlab_level}{levels of x lab}

\item{fill_level}{levels of fill (legend)}

\item{jitter}{if adding jitter points, default T}

\item{fill.col}{if filling the bars with colors, default T}

\item{fill.alpha}{the alpha of fill colors, default 1}

\item{title}{plot title, default NULL}

\item{x_lab}{plot x lab, default NULL}

\item{y_lab}{plot y lab, default ''}

\item{point.size}{the size of the jitter points, default 3}

\item{title.text.size}{text size of the plot title, default 20}

\item{xlab.text.size}{text size of the x lab, default 18}

\item{ylab.text.size}{text size of the y lab, default 18}

\item{xtitle.text.size}{text size of the x title, default 18}

\item{ytitle.text.size}{text size of the y title, , default 18}

\item{xlab.angle}{if rotate the x labs, default 0}

\item{legend.position}{legend position default top}

\item{legend.text.size}{text size of the legend , default 12}

\item{pal}{color for filling  default pal_lancet}

\item{plot.theme}{theme for plot, deault prism}
}
\value{
ggplot plot
}
\description{
Barplot (dodge, adding points and prism theme) with data presented as long, do not analysis and provide mean or sd or counts. Suitable for y data type of continuous!
}
\examples{
barplot_jitter_prism(data = tmp, x = 'Group', value = 'Value', fill = 'Cell', fill.col = F, jitter = F)
}
