% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplot_percent_fill}
\alias{barplot_percent_fill}
\title{Barplot (fill position showing percent in the group) with data presented as long, do not analysis and provide mean or sd or counts. Suitable for xtab counting results presenting!}
\usage{
barplot_percent_fill(
  data,
  x,
  count,
  na.rm = F,
  lab_text = T,
  lab.text.size = 4,
  lab.text.color = "black",
  xlab_level,
  bar_width = 0.9,
  title = NULL,
  x_lab = NULL,
  y_lab = "",
  base.size = 18,
  legend.position = "top",
  legend.size = 12,
  xlab.angle = 0,
  pal = (ggsci::pal_lancet(palette = c("lanonc"), alpha = 0.6))(9),
  alpha.bar = 1,
  chinese = F,
  plot.theme
)
}
\arguments{
\item{data}{Dataframe with long form (raw data, do not analysis!); all discrete data}

\item{x}{x column name be counted , data type must be discrete}

\item{count}{y column name to be counted , data type must be discrete}

\item{na.rm}{if remove NA values, default F}

\item{lab_text}{if show labs of percent in the group , default T}

\item{lab.text.size}{the text size of the labs, default 4}

\item{lab.text.color}{the text colors of the labs, default black}

\item{xlab_level}{levels of x lab}

\item{bar_width}{bar width default 0.9}

\item{title}{plot title}

\item{x_lab}{plot x lab title}

\item{y_lab}{plot y lab title}

\item{base.size}{base text size of the theme default 18}

\item{legend.position}{legend position default top}

\item{legend.size}{legend text size default 12}

\item{xlab.angle}{if rotate the x labs defualt 0}

\item{pal}{color for filling  default pal_lancet}

\item{alpha.bar}{alpha of the bar, default 1}

\item{chinese}{if text in Chinese ,default False}

\item{plot.theme}{theme for plot deault classic}
}
\value{
ggplot plot
}
\description{
Barplot (fill position showing percent in the group) with data presented as long, do not analysis and provide mean or sd or counts. Suitable for xtab counting results presenting!
}
\examples{
barplot_percent_fill(data = sur_data, x = 'paper_BRCA_Subtype_PAM50', count = 'tmb', na.rm = T)
}
