% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot}
\alias{boxplot}
\title{Boxplot with data presented as long, do not analysis and provide mean or sd or counts. Suitable for y data type of continuous!}
\usage{
boxplot(
  data,
  x_value,
  y_value,
  group,
  paired,
  box.fill = T,
  box.color = F,
  jitter = F,
  xlab_level,
  group_level,
  title = NULL,
  x_lab = NULL,
  y_lab = "",
  legend.title = "",
  base.size = 18,
  legend.position = "top",
  legend.size = 12,
  xlab.angle = 0,
  pal = (ggsci::pal_lancet(palette = c("lanonc"), alpha = 0.6))(9),
  alpha.box = 0.8,
  add.p = F,
  method = "t.test",
  chinese = F,
  plot.theme
)
}
\arguments{
\item{data}{dataframe with long form (raw data, do not analysis!); y continuous data}

\item{x_value}{x column name , data type must be discrete}

\item{y_value}{Y column name,  data type must be continuous for mean/sd}

\item{group}{group column name, data type must be discrete (for filling sub type of the x value)}

\item{paired}{if paired, give a column name for paired information}

\item{box.fill}{if fill the box with palette ,default T ,if false fill will be white}

\item{box.color}{if give the box line color with palette, default F, line will be black}

\item{jitter}{if add jitter point on the box}

\item{xlab_level}{levels of x lab}

\item{group_level}{levels of group (legend)}

\item{title}{plot title}

\item{x_lab}{plot x lab}

\item{y_lab}{plot y lab}

\item{legend.title}{plot legend title}

\item{base.size}{base text size of the theme default 18}

\item{legend.position}{legend position default top}

\item{legend.size}{legend text size default 12}

\item{xlab.angle}{if rotate the x labs defualt 0}

\item{pal}{color for filling  default pal_lancet}

\item{alpha.box}{alpha of the box filling, default 0.8}

\item{chinese}{if text in Chinese ,default False}

\item{plot.theme}{plot.theme}
}
\value{
ggplot plot object
}
\description{
Boxplot with data presented as long, do not analysis and provide mean or sd or counts. Suitable for y data type of continuous!
}
\examples{
boxplot(data = ToothGrowth, x_value = 'dose', y_value = 'len', group = 'supp', jitter = T , box.fill = T, box.color = T, xlab_level = c("2", "1", "0.5"), plot.theme = ggprism::theme_prism())
          boxplot(data = data, x_value = 'group', y_value = 'Gene_Expression', group = 'group', title = '', y_lab = 'Gene Expression')
          boxplot(data = data, x_value = 'group', y_value = 'Gene_Expression', group = 'group', title = '', y_lab = 'Gene Expression', paired = 'pair')
}
