% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationplot.R
\name{cor.analysis}
\alias{cor.analysis}
\title{Calculate the cor coefficient with 95\%CI from two factors analysed using x, y var names}
\usage{
cor.analysis(data, factor1, factor2, method = "spearman")
}
\arguments{
\item{data}{Raw data frame with row of samples, colunames of multiple factors}

\item{factor1}{column names of x var name}

\item{factor2}{column names of y var name}

\item{method}{analysis methods can be "pearson", "spearman", "kendall", default spearman}
}
\value{
a vector of factor1, factor2, method, cor coefficient, p value, cor 95\%lower, cor 95\%upper
}
\description{
Calculate the cor coefficient with 95\%CI from two factors analysed using x, y var names
}
\examples{
cor.analysis(data = dat, gene1 = "CD8A", gene2 = 'CXCL9', method = 'spearman')
}
