% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationplot.R
\name{cor_95CI}
\alias{cor_95CI}
\title{calculate the 95\% confidence interval of cor coefficient}
\usage{
cor_95CI(cor.estimate, total.sample)
}
\arguments{
\item{cor.estimate}{cor coefficient analysed from cor.test function}

\item{total.sample}{number of total samples used}
}
\value{
a vector of  95\% lower and upper  confidence of cor coefficient
}
\description{
calculate the 95\% confidence interval of cor coefficient
}
\examples{
cor_95CI(cor.estimate = dd$estimate, total.sample = nrow(data))
}
