% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationplot.R
\name{cor_buble_plot}
\alias{cor_buble_plot}
\title{Bubble plot showing correlation of two different matrix,  bubble size for -log10(p), bubble color of correlation r}
\usage{
cor_buble_plot(
  df_row,
  df_column,
  cor.method = "spearman",
  pal = c("#4d685c", "white", "#a84b7c"),
  x.lab.size = 14,
  x.lab.angle = 90,
  y.lab.size = 14,
  y.lab.angle = 0,
  legend.title.size = 14,
  legend.text.size = 12,
  bulb.size.limit = c(0, 200),
  bulb.size.break = seq(0, 200, 50),
  bulb.size.labels,
  bulb.color.limit = c(-1, 1),
  bulb.color.midpoint = 0,
  bulb.color.break = seq(-1, 1, 0.5)
)
}
\arguments{
\item{df_row}{dataframe with column names on the row of the correlation plot}

\item{df_column}{dataframe with column names on the column of the correlation plot}

\item{cor.method}{methods for correlation analysis, default spearman}

\item{pal}{color for the bubble , a vector of 3 color numbers c(low, mid, high),  default c("#4d685c", "white", "#a84b7c")}

\item{x.lab.size}{text size of the row name of the heatmap; default 14}

\item{x.lab.angle}{text angle of the row name of the heatmap; default 90}

\item{y.lab.size}{text size of the column name of the heatmap; default 14}

\item{y.lab.angle}{text angle of the column names of the heatmap; default 0}

\item{legend.title.size}{text size of the legend title, default 14}

\item{legend.text.size}{text size of the legend text, default 12}

\item{bulb.size.limit}{size limit  for the bubble (-log10P) , a vector of 2 numbers c(low, high),  default c(0, 200)}

\item{bulb.size.break}{breaks of the bubble size, a vector of numbers in the limits, default seq(0, 200, 50)}

\item{bulb.size.labels}{labels of the bubble size, a vector of numbers in the limits, default same with bulb.size.break}

\item{bulb.color.limit}{color limit of the bubble (r),  a vector of 2 numbers c(low, high), default c(-1, 1)}

\item{bulb.color.midpoint}{middle point of the bubble color , default 0}

\item{bulb.color.break}{breaks of the bubble color, a vector of numbers in the limits, default seq(-1,1,0.5)}
}
\value{
a ggplot object of bubble plot showing correlation of two different matrix
}
\description{
Bubble plot showing correlation of two different matrix,  bubble size for -log10(p), bubble color of correlation r
}
\examples{
cor_buble_plot(df_row = dat2, df_column =  dat1, bulb.color.limit = c(-0.2, 0.8), bulb.color.midpoint = 0.4, bulb.color.break = seq(-0.2,0.8,0.2),
          x.lab.size = 12, x.lab.angle = 75, y.lab.size = 12)
}
