% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationplot.R
\name{cor_heatmap_plot2}
\alias{cor_heatmap_plot2}
\title{Heatmap showing correlation of two different matrix,  could show p values with/without significance (P < 0.05)}
\usage{
cor_heatmap_plot2(
  df_row,
  df_column,
  cor.method = "spearman",
  pal = c("#0074b3", "white", "#982b2b"),
  x.lab.size = 14,
  x.lab.angle = 90,
  y.lab.size = 14,
  y.lab.angle = 0,
  legend.title.size = 14,
  legend.text.size = 12,
  legend.limit = c(-1, 1),
  legend.midpoint = 0,
  legend.break = seq(-1, 1, 0.5)
)
}
\arguments{
\item{df_row}{dataframe with column names on the row of the correlation heatmap}

\item{df_column}{dataframe with column names on the column of the correlation heatmap}

\item{cor.method}{methods for correlation analysis, default spearman}

\item{pal}{color for the heatmap , a vector of 3 color numbers c(low, mid, high),  default c("#0074b3", "white", "#982b2b")}

\item{x.lab.size}{text size of the row name of the heatmap; default 14}

\item{x.lab.angle}{text angle of the row name of the heatmap; default 90}

\item{y.lab.size}{text size of the column name of the heatmap; default 14}

\item{y.lab.angle}{text angle of the column names of the heatmap; default 0}

\item{legend.title.size}{text size of the legend title, default 14}

\item{legend.text.size}{text size of the legend text, default 12}

\item{legend.limit}{limit of the the legend bar,  a vector of 2 numbers c(low, high), default c(-1, 1)}

\item{legend.midpoint}{middle point of the legend bar , default 0}

\item{legend.break}{breaks of the legend bar, a vector of numbers in the limits, default seq(-1,1,0.5)}
}
\value{
a ggplot object of heatmap showing correlation of two different matrix
}
\description{
Heatmap showing correlation of two different matrix,  could show p values with/without significance (P < 0.05)
}
\examples{
cor_heatmap_plot2(df_row = dat2, df_column = dat1, cor.method = 'spearman',
          legend.limit = c(-0.2, 1), legend.midpoint = 0.4, legend.break = seq(-0.2, 1, 0.4), pal = c("#4d685c", "white", "#a84b7c"))
}
