% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationplot.R
\name{cor_link_plot}
\alias{cor_link_plot}
\title{corelation showing on two different matrix, not all the correlation of factors in one row}
\usage{
cor_link_plot(df_matrix, df_link, text.size = 2.5)
}
\arguments{
\item{df_matrix}{correlation show using matrix}

\item{df_link}{correlation show using link lines on the factors of the matrix}

\item{text.size}{text size ,default 2.5}
}
\value{
a link line corplot
}
\description{
corelation showing on two different matrix, not all the correlation of factors in one row
}
\examples{
cor_link_plot(df_matrix = df_expr, df_link = immucell_df)
}
