% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationplot.R
\name{cor_matrix2}
\alias{cor_matrix2}
\title{show correlation plot for multiple variables (more than two vars, show by matrix, using ggcor package)}
\usage{
cor_matrix2(cor.matrix, shape = "square", text.size = 2.5)
}
\arguments{
\item{cor.matrix}{data frame with different vars: row samples, columns different var names (cor analysis in function)}

\item{shape}{can be 'circle', 'square'; default square}

\item{text.size}{text size default 2.5}
}
\value{
a cor plot for multi factors
}
\description{
show correlation plot for multiple variables (more than two vars, show by matrix, using ggcor package)
}
\examples{
cor_matrix2(cor.matrix = df_expr)
}
