% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationplot.R
\name{corrplot_Margin}
\alias{corrplot_Margin}
\title{A cor plot with margin plots analysed using x, y var names}
\usage{
corrplot_Margin(
  data,
  x_value,
  y_value,
  log.Data = "none",
  cor.methods = "spearman",
  point.size = 2,
  point.color = "#025259",
  text.position = "topright",
  text.size = 5,
  text.color = "red",
  Marginal.type = "histogram",
  Marginal.color = "black",
  Marginal.fill = "#d94f04",
  title = NULL,
  x_lab,
  y_lab,
  base.size = 18,
  legend.position = "top",
  legend.size = 12,
  plot.theme
)
}
\arguments{
\item{data}{dataframe with long form; row samples, columns different varnames, such as genes, data type numeric/continuous data}

\item{x_value}{varname of x ,such as gene1}

\item{y_value}{varname of y ,such as gene2}

\item{log.Data}{if log the data, can be none, log2, log10}

\item{cor.methods}{cor analysis methods: pearson, kendall, spearman, default spearman}

\item{point.size}{size of the points,default 2}

\item{point.color}{color of the points, default "#025259"}

\item{text.position}{text position of the cor/p value, can be topright, topleft, bottomright, bottomleft, none; default topright}

\item{text.size}{text size ,default 5}

\item{text.color}{text color ,default red}

\item{Marginal.type}{type of Marginal plot, can be histogram, density, boxplot, densigram, none; default histogram}

\item{Marginal.color}{color of Marginal plot lines, defualt black}

\item{Marginal.fill}{color of Marginal plot fillings, defualt '#d94f04'}

\item{title}{plot title}

\item{x_lab}{plot x lab}

\item{y_lab}{plot y lab}

\item{base.size}{base text size of the theme default 18}

\item{legend.position}{legend position default top}

\item{legend.size}{legend text size default 12}

\item{plot.theme}{plot.theme}
}
\value{
ggplot plot object
}
\description{
A cor plot with margin plots analysed using x, y var names
}
\examples{
corrplot_Margin(data = dat, x_value = 'LATS2', y_value = 'ALKBH5', base.size = 20)
          corrplot_Margin(data = dat, x_value = 'LATS2', y_value = 'ALKBH5', log.Data = 'log10', text.position = 'bottomright', Marginal.type = 'densigram', title = 'corplot')
}
