% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationplot.R
\name{corrplot_Rug}
\alias{corrplot_Rug}
\title{A cor plot with rug analysed using x, y var names}
\usage{
corrplot_Rug(
  data,
  x_value,
  y_value,
  log.Data = "none",
  cor.methods = "spearman",
  point.size = 1,
  point.color = "#984ea3",
  smoothline.color = "#fdc086",
  rug.color = "#7fc97f",
  text.size = 5,
  title = NULL,
  x_lab,
  y_lab,
  base.size = 15,
  legend.position = "top",
  legend.size = 12,
  plot.theme
)
}
\arguments{
\item{data}{dataframe with long form; row samples, columns different varnames, such as genes, data type numeric/continuous data}

\item{x_value}{varname of x ,such as gene1}

\item{y_value}{varname of y ,such as gene2}

\item{log.Data}{if log the data, can be none, log2, log10}

\item{cor.methods}{cor analysis methods: pearson, kendall, spearman, default spearman}

\item{point.size}{size of the points,default 1}

\item{point.color}{color of the points, default "#984ea3"}

\item{smoothline.color}{color of the smoothline, default "#fdc086"}

\item{rug.color}{color of the "#7fc97f"}

\item{text.size}{size of the text(cor/p value), default 5}

\item{title}{plot title}

\item{x_lab}{plot x lab}

\item{y_lab}{plot y lab}

\item{base.size}{base text size of the theme default 15}

\item{legend.position}{legend position default top}

\item{legend.size}{legend text size default 12}

\item{plot.theme}{plot.theme}
}
\value{
ggplot plot object
}
\description{
A cor plot with rug analysed using x, y var names
}
\examples{
corrplot_Rug(data = sur_data, x_value = 'ASNS', y_value = 'ATF4', text.size = 5, base.size = 20)
}
