% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestplot.R
\name{forestplot_identity}
\alias{forestplot_identity}
\title{Title}
\usage{
forestplot_identity(
  data,
  x_value,
  y_value,
  y_min,
  y_max,
  group,
  xlab_level,
  group_level,
  position.dodge = 0.3,
  point.shape = "circle",
  point.size = 3,
  line.width = 0.5,
  errbar.width = 0,
  y.limits,
  y.breaks,
  coord.filp = T,
  title = NULL,
  x_lab = NULL,
  y_lab = "",
  base.size = 18,
  legend.position = "top",
  legend.size = 12,
  pal = (ggsci::pal_lancet(palette = c("lanonc"), alpha = 0.6))(9),
  plot.theme
)
}
\arguments{
\item{data}{data frame anlysised previously with y value and its 95\% CI intervals}

\item{x_value}{column name of x value in the dataframe}

\item{y_value}{column name of y value in the dataframe}

\item{y_min}{column name of y value min (95\% low) in the dataframe}

\item{y_max}{column name of y value max (95\% upper) in the dataframe}

\item{group}{column name of group in the dataframe}

\item{xlab_level}{levels of x lab}

\item{group_level}{levels of group(legends)}

\item{position.dodge}{position dodge of errbor/points default 0.3}

\item{point.shape}{shape of the points, can be circle/square, default circle}

\item{point.size}{size of the points, default 3}

\item{line.width}{width of the errbar line, default 0.3}

\item{errbar.width}{length of th errbar, default 0 (No line observed)}

\item{y.limits}{limits of y axis (x axis if filped), a vector: c(min, max), such as: c(0, 1)}

\item{y.breaks}{breaks of y axis (x axis if filped), a vector:  seq(min, max, interval), such as: seq(0, 1, 0.2). Or design values you wish}

\item{coord.filp}{if filp the plot,default T}

\item{title}{plot title}

\item{x_lab}{plot x lab}

\item{y_lab}{plot y lab}

\item{base.size}{base text size of the theme default 18}

\item{legend.position}{legend position default top}

\item{legend.size}{legend text size default 12}

\item{pal}{color for filling  default pal_lancet}

\item{plot.theme}{theme for plot deault classic}
}
\value{
a ggplot2 object can add geom_hline as you wish
}
\description{
Title
}
\examples{
forestplot_identity()
}
