% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pieplot.R
\name{ggpie_identity}
\alias{ggpie_identity}
\title{pie plot using ggpubr::ggpie, data given raw counts / calculated percentage (a number < 1; e.g. 0.55)
suitable for few groups, labs inside of the pie}
\usage{
ggpie_identity(
  data,
  group,
  value,
  labs,
  group.levels,
  pal = (ggsci::pal_futurama(alpha = 0.8))(12),
  chinese = F,
  lab.size = 14,
  legend.position = "top",
  legend.size = 12
)
}
\arguments{
\item{data}{dataframe with a column of group(as lab names), at least one colunm of values: can be a colunm of counts or a column with percent,
labs self-defined can also be used}

\item{group}{colunm name of group for legends}

\item{value}{colunm name for counts/percentage}

\item{labs}{colunm name for labs shown on the pie(inside), default percentage}

\item{group.levels}{define group order in the legend}

\item{pal}{color of the pie section}

\item{chinese}{if the text is Chinese, if T font faminly change to STHeitiSC-Medium, default F}

\item{lab.size}{text size on the lab}

\item{legend.position}{position of the legend}

\item{legend.size}{text size on the legend}
}
\value{
a ggplot pie plot with labs in the pie
}
\description{
pie plot using ggpubr::ggpie, data given raw counts / calculated percentage (a number < 1; e.g. 0.55)
suitable for few groups, labs inside of the pie
}
\examples{
myplotting::ggpie_identity(data = df, group = 'group', value = 'value', labs = 'labs')
}
