% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pieplot.R
\name{ggpie_stat}
\alias{ggpie_stat}
\title{pie plot using ggpie::ggpie, data given as long formation, plot show statistic result. don't count the data for preparing.}
\usage{
ggpie_stat(
  data,
  group_key,
  label_info = "all",
  label_rotate = "circle",
  label_split = " ",
  label_size = 4,
  label_pos = "in",
  label_threshold = NULL,
  pal = (ggsci::pal_futurama(alpha = 0.8))(12),
  label_color = "black",
  border_color = "black",
  border_size = 0.5,
  chinese = F,
  title = NULL,
  title.size = 14,
  legend.position = "top",
  legend.size = 10
)
}
\arguments{
\item{data}{dataframe with at least one column of group (discrete data) , ggpie will give count and ratio by itself.}

\item{group_key}{column name of the group}

\item{label_info}{Label information type, chosen from count, ratio and all (count and ratio). Default: all}

\item{label_rotate}{Label style, chosen from circle, horizon and none (no label). Default: circle.}

\item{label_split}{Pattern used to split the label, support regular expression. Default: ' ', give a blank space ' ' could make the labels into two lines}

\item{label_size}{Size of the label. Default: 4.}

\item{label_pos}{Label position, chosen from in and out. Default: in}

\item{label_threshold}{Threshold of the ratio to determine label position (in/out pie). Default: NULL.}

\item{pal}{Colors used. Default: ggsci::pal_futurama(alpha = 0.8)(12)}

\item{label_color}{Color of the label. Default: black.}

\item{border_color}{Border color. Default: black.}

\item{border_size}{Border thickness. Default: 0.5.}

\item{chinese}{if use Chinese text in group levels}

\item{title}{plot title Default:NULL}

\item{title.size}{text size of title ,default 14}

\item{legend.position}{position of legend , default top}

\item{legend.size}{text size of legend ,default 10}
}
\value{
a ggplot pie plot
}
\description{
pie plot using ggpie::ggpie, data given as long formation, plot show statistic result. don't count the data for preparing.
}
\examples{
ggpie_stat(data = diamonds, group_key = "cut", label_info = "all", label_rotate = "circle", label_size = 3, label_pos = "in")
}
