% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap_ggplot2}
\alias{heatmap_ggplot2}
\title{Title: heatmap plotting with ggplot2, matrix could be discrete / continues values}
\usage{
heatmap_ggplot2(
  df,
  row.factor.levels,
  col.factor.levels,
  fill.factor.levels,
  pal.discrete = (ggsci::pal_lancet(palette = c("lanonc"), alpha = 0.6))(9),
  pal.continue = c("#498EA4", "white", "#E54924"),
  pal.continue.limit = NULL,
  pal.continue.break = NULL,
  border.color = "grey",
  legend.title = "",
  legend.title.size = 18,
  legend.text.size = 16,
  legend.position = "right",
  show.rowNames = T,
  row.text.size = 16,
  row.text.angle = 0,
  show.colNames = T,
  col.text.size = 16,
  col.text.angle = 90
)
}
\arguments{
\item{df}{a dataframe for heat map, row samples column factors}

\item{row.factor.levels}{row name orders ranked as row.factor.levels}

\item{col.factor.levels}{col name orders ranked as col.factor.levels}

\item{fill.factor.levels}{legend factor ranked as fill.factor.levels if filling for discrete values}

\item{pal.discrete}{discrete pal color pal for heatmap, default ggsci::pal_lancet(palette = c("lanonc"), alpha = 0.6)(9)}

\item{pal.continue}{continue pal color pal for heatmap, could be two or three, default c('#498EA4', 'white', '#E54924')}

\item{pal.continue.limit}{limits for continue pal color pal, default max and min value}

\item{pal.continue.break}{breaks for continue pal color pal, default a seq vector cutting the bar average}

\item{border.color}{border line color of the tile,  default grey}

\item{legend.title}{legend title of the heatmap bar,  default blank ''}

\item{legend.title.size}{text size of the legend title, default 18}

\item{legend.text.size}{text size of the legend lab, default 16}

\item{legend.position}{legend bar position ,default right, 'none' do not show the bar}

\item{show.rowNames}{if show row names default T}

\item{row.text.size}{text size of the row names, default 16}

\item{row.text.angle}{text angle of the row names, default 0}

\item{show.colNames}{if show column names default T}

\item{col.text.size}{text size of the col names, default 16}

\item{col.text.angle}{text angle of the column names, default 90}
}
\value{
a ggplot2 object of heatmap (the main plot)
}
\description{
Title: heatmap plotting with ggplot2, matrix could be discrete / continues values
}
\examples{
heatmap_ggplot2(df = df_wider, col.factor.levels = c('low', 'high'), pal.continue = c("#0074b3", "#982b2b"), legend.title = 'gene expression (TPM)',
          pal.continue.limit = c(0, 100), pal.continue.break = c( 20, 40, 60, 80))

}
