% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineplot.R
\name{lineplot_identity}
\alias{lineplot_identity}
\title{lineplot with data presented identity (as rawdata give mean , sd)}
\usage{
lineplot_identity(
  data,
  x_value,
  y_value,
  group,
  errbar,
  lab_text = F,
  shape.size = 3,
  line.width = 0.7,
  text.size = 3,
  xlab_level,
  group_level,
  title = NULL,
  x_lab = NULL,
  y_lab = "",
  base.size = 15,
  legend.position = "top",
  legend.size = 12,
  xlab.angle = 0,
  pal = (ggsci::pal_lancet(palette = c("lanonc"), alpha = 0.6))(9),
  chinese = F,
  plot.theme
)
}
\arguments{
\item{data}{data frame anlysised previously with mean, sd}

\item{x_value}{column name of x value in the dataframe}

\item{y_value}{column name of y value in the dataframe, such as mean}

\item{group}{column name of group for line colors}

\item{errbar}{column name of error bar}

\item{lab_text}{adding text of y value on the point, default False}

\item{shape.size}{size of the points(shape), default 3}

\item{line.width}{width of the lines, default 0.7}

\item{text.size}{size of the text add on the shapes, default 3}

\item{xlab_level}{levels of x lab}

\item{group_level}{levels of group lab}

\item{title}{plot title}

\item{x_lab}{plot x lab}

\item{y_lab}{plot y lab}

\item{base.size}{base text size of the theme default 15}

\item{legend.position}{legend position default top}

\item{legend.size}{legend text size default 12}

\item{xlab.angle}{if rotate the x labs defualt 0}

\item{pal}{color for filling  default pal_lancet}

\item{chinese}{if text in Chinese ,default False}

\item{plot.theme}{plot.theme}
}
\value{
ggplot plot object
}
\description{
lineplot with data presented identity (as rawdata give mean , sd)
}
\examples{
lineplot_identity(data, x, y )
}
