% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoplot.R
\name{paretoplot_axis}
\alias{paretoplot_axis}
\title{A special bar plot with ordered bars based on y value, and lab show on the x axis
color (< 4) highlight the group on the axis by a filling group with factor data}
\usage{
paretoplot_axis(
  data,
  x_value,
  y_value,
  fill_group,
  fill_level,
  descend = F,
  bar.width = 0.6,
  x.lab.size = 10,
  pal = (ggsci::pal_lancet(palette = c("lanonc"), alpha = 0.6))(9),
  title = NULL,
  x_lab = NULL,
  y_lab = "",
  base.size = 15,
  legend.position = "top",
  legend.size = 12,
  chinese = F,
  cord.filp = F
)
}
\arguments{
\item{data}{a dataframe with  x lab names (also shown as text); y value (numeric); group for filling colors(not neccisdary, if not provided, bars with only one color)}

\item{x_value}{column name of x value in the dataframe}

\item{y_value}{column name of y value in the dataframe, such as mean}

\item{fill_group}{column name of filling for subtype (if not provided, using the first color in the pallette)}

\item{fill_level}{levels of fill lab}

\item{descend}{if using desceding order of y}

\item{bar.width}{bar width default 0.6}

\item{x.lab.size}{text size on the x axis default 10}

\item{pal}{color for filling  default pal_lancet}

\item{title}{plot title default NULL}

\item{x_lab}{plot x lab default NULL}

\item{y_lab}{plot y lab default ''}

\item{base.size}{base text size of the theme default 15}

\item{legend.position}{legend position default top}

\item{legend.size}{legend text size default 12}

\item{chinese}{if text in Chinese ,default False}

\item{cord.filp}{if filp the plot,default F}
}
\value{
a ggplot barplot with order y values ; labs outside the axis with filling color (less than 4)
}
\description{
A special bar plot with ordered bars based on y value, and lab show on the x axis
color (< 4) highlight the group on the axis by a filling group with factor data
}
\examples{
paretoplot_axis(data = plot_data, x_value = 'cell_line', y_value = 'Expression', fill_group = 'pick_color')
          paretoplot_axis(data = plot_data, x_value = 'cell_line', y_value = 'Expression')

}
