% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoplot.R
\name{paretoplot_lab}
\alias{paretoplot_lab}
\title{A special bar plot with ordered bars based on y value, and lab show on the bars
the position of labs based on y value}
\usage{
paretoplot_lab(
  data,
  x_value,
  y_value,
  fill_group,
  fill_level,
  descend = F,
  bar.width = 0.6,
  lab.size = 3,
  lab.position = 0.1,
  pal = (ggsci::pal_lancet(palette = c("lanonc"), alpha = 0.6))(9),
  title = NULL,
  x_lab = NULL,
  y_lab = "",
  base.size = 18,
  legend.position = "top",
  legend.size = 12,
  chinese = F,
  cord.filp = F
)
}
\arguments{
\item{data}{a dataframe with  x lab names (also shown as text); y value (numeric); group for filling colors(not neccisdary, if not provided, bars with only one color)}

\item{x_value}{column name of x value in the dataframe}

\item{y_value}{column name of y value in the dataframe, such as mean}

\item{fill_group}{column name of filling for subtype (if not provided, using the first color in the pallette)}

\item{fill_level}{levels of fill lab}

\item{descend}{if using desceding order of y}

\item{bar.width}{bar width default 0.6}

\item{lab.size}{lab text size default 3}

\item{lab.position}{the start position of labs (y value), default 0.1}

\item{pal}{color for filling  default pal_lancet}

\item{title}{plot title default NULL}

\item{x_lab}{plot x lab default NULL}

\item{y_lab}{plot x lab default ''}

\item{base.size}{base text size of the theme default 18}

\item{legend.position}{legend position default top}

\item{legend.size}{legend text size default 12}

\item{chinese}{if text in Chinese ,default False}

\item{cord.filp}{if filp the plot,default F}
}
\value{
a ggplot barplot with order y values ; labs outside the bar(2 sides based on y value, like gsea plot)
}
\description{
A special bar plot with ordered bars based on y value, and lab show on the bars
the position of labs based on y value
}
\examples{
paretoplot(data = plot_data, x_value = 'cell_line', y_value = 'gene_effect', fill_group = 'pick_color', fill_level = c('TRUE', 'FALSE'), descend = T, legend.size = 8)
          paretoplot1(data = plot_data, x_value = 'cell_line', y_value = 'gene_effect',  descend = F)

}
