% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoplot.R
\name{paretoplot_neighbor}
\alias{paretoplot_neighbor}
\title{A special bar plot with ordered bars based on y value, and lab show beside the bar top (like a neighborhood)
can give a lab column you want, not have to be the x lab value}
\usage{
paretoplot_neighbor(
  data,
  x_value,
  y_value,
  fill,
  add.labs = T,
  labs,
  show.xlab = T,
  fill_level,
  lab.size = 3,
  descend = F,
  y.limit,
  bar_width = 0.6,
  pal = (ggsci::pal_lancet(palette = c("lanonc"), alpha = 0.6))(9),
  title = NULL,
  x_lab = NULL,
  y_lab = "",
  base.size = 18,
  legend.position = "top",
  legend.size = 12,
  chinese = F,
  cord.filp = F
)
}
\arguments{
\item{data}{a dataframe with  x value; y value (numeric); group for filling colors(not neccisdary, if not provided, bars with only one color); lab column, not neccissary}

\item{x_value}{column name of x value in the dataframe}

\item{y_value}{column name of y value in the dataframe, such as mean}

\item{fill}{column name of filling for subtype (if not provided, using the first color in the pallette)}

\item{add.labs}{if add labs colunm on the bars}

\item{labs}{column name of labs in the dataframe, if not provided, using x value instead}

\item{show.xlab}{if show x lab text; if labs is the same, do not have to show x labs.}

\item{fill_level}{levels of fill lab}

\item{lab.size}{lab text size default 3}

\item{descend}{if using desceding order of y}

\item{y.limit}{y.limit need a vector with 2 numbers as the ymin and ymax: e.g y.limit = c(-30, 40)}

\item{bar_width}{bar width default 0.6}

\item{pal}{color for filling  default pal_lancet}

\item{title}{plot title default NULL}

\item{x_lab}{plot x lab default NULL}

\item{y_lab}{plot y lab default ''}

\item{base.size}{base text size of the theme default 18}

\item{legend.position}{legend position default top}

\item{legend.size}{legend text size default 12}

\item{chinese}{if text in Chinese ,default False}

\item{cord.filp}{if filp the plot,default F}
}
\value{
a ggplot barplot with order y values; labs beside the bar (like adding text on the bar)
}
\description{
A special bar plot with ordered bars based on y value, and lab show beside the bar top (like a neighborhood)
can give a lab column you want, not have to be the x lab value
}
\examples{
paretoplot_neighbor(data = plot_data, x_value = 'cell_line', y_value = 'Expression', fill = 'pick_color')
}
