% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{pheatmap_plotting}
\alias{pheatmap_plotting}
\title{heatmap plotting with pheatmap package}
\usage{
pheatmap_plotting(
  df,
  limitMax,
  limitMin,
  scale = F,
  show.rowName = T,
  show.colName = T,
  row.cluster = F,
  col.cluster = F,
  cellwidth = 50,
  cellheight = 20,
  pal = c("navy", "white", "firebrick3"),
  color_levelnum = 50,
  border.color = NA,
  col.angle = 90,
  rowtext.size = 14,
  coltext.size = 18,
  adding.legend = T,
  legend.breaks = NA,
  legend.labels = NA,
  ac_df = NA,
  ann_colors = NA,
  add.rowgaps = c(0),
  add.colgaps = c(0)
)
}
\arguments{
\item{df}{a dataframe or matrix for heat map, row samples column factors}

\item{limitMax}{limit of the max value of the matrix, avoid too large value}

\item{limitMin}{limit of the max value of the matrix, avoid too small value}

\item{scale}{if scaleing the matrix, can also use log before heatmap plotting, default F}

\item{show.rowName}{if show row Name, default T}

\item{show.colName}{if show column Name, default T}

\item{row.cluster}{if clust row, default F}

\item{col.cluster}{if clust column, default F}

\item{cellwidth}{cell width, default 50}

\item{cellheight}{cell height, default 20}

\item{pal}{color pal for heatmap  a vector of three color values, default  c("navy", "white", "firebrick3")}

\item{color_levelnum}{levels setting for color change, default 50}

\item{border.color}{border line color, default NA, without border line color;}

\item{col.angle}{angle of column name ,can be 270、0、45、90、315, default 90}

\item{rowtext.size}{rowname text size , default 14}

\item{coltext.size}{colname text size , default 18}

\item{adding.legend}{if adding the legend figure for heatmap, default T}

\item{legend.breaks}{a vectors of numeric to divide plot values}

\item{legend.labels}{a vectors of strings to label legend.breaks, must be consistant with break cut values}

\item{ac_df}{annotation dataframe for colunms}

\item{ann_colors}{a color list for annotation dataframes , could be more than one palette
e.g., ann_colors = list(group1= c("white", "tomato4"),  group2 = c(X1 = "slateblue3", X2 = "red2"), group3 = c(phylum1 = "#7D26CD", phylum2 = "#E7298A", phylum3 = "#66A61E"))}

\item{add.rowgaps}{gaps split the row, a numeric vector with the beginning row numbers to split}

\item{add.colgaps}{gaps split the column, a numeric vector with the beginning colnumn numbers to split}
}
\value{
a grid object of heatmap plot
}
\description{
heatmap plotting with pheatmap package
}
\examples{
myplotting::pheatmap_plotting(df = n, cellheight = 2, show.rowName = F, coltext.size = 10, ac_df = ac, add.rowgaps = c(50, 80), add.colgaps = c(3))
}
