% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pieplot.R
\name{pieplot_base}
\alias{pieplot_base}
\title{pie plot using base R, data given raw counts / calculated percentage (a number < 1; e.g. 0.55)
suitable for more groups, labs outside of the pie}
\usage{
pieplot_base(
  data,
  radius = 0.6,
  Value.Count,
  Value.Prob,
  group.lab,
  add_group.lab_in_pie = T,
  title = "",
  lab.size = 0.8,
  legend.postion = "topleft",
  legend.size = 0.8,
  pal = (ggsci::pal_futurama(alpha = 0.8))(12),
  mar = c(1, 1, 1, 1),
  family = "Arial",
  Chinese = F
)
}
\arguments{
\item{data}{dataframe with a column of group(as lab names), at least one colunm of values: can be  a colunm of counts or a column with percent}

\item{radius}{size of pie ,default 0.6}

\item{Value.Count}{colunm name for counts}

\item{Value.Prob}{colunm name for percentage (must be number < 1)}

\item{group.lab}{group labs of percentage (also could be shown on legend)}

\item{add_group.lab_in_pie}{if add the lab with percentage on the pie (defalut T)}

\item{title}{title of the plot}

\item{lab.size}{text size on the lab}

\item{legend.postion}{legend position(defalut topleft), if none ,no legend added ; could be one of “bottomright”, “bottom”, “bottomleft”, “left”, “topleft”, “top”, “topright”, “right”, “center”}

\item{legend.size}{text size on the legend}

\item{pal}{color of the pie section}

\item{mar}{mar defautl c(1,1,1,1)}

\item{family}{family default Arial}

\item{Chinese}{if the text is Chinese, if T font faminly change to STHeitiSC-Medium, default F}
}
\value{
a base R pie plot with labs outside the pie
}
\description{
pie plot using base R, data given raw counts / calculated percentage (a number < 1; e.g. 0.55)
suitable for more groups, labs outside of the pie
}
\examples{
pieplot_base(data = df, Value.Count = 'growth', group.lab = 'region')
          pieplot_base(data = df, Value.Prob = 'prob', group.lab = 'region', radius = 0.8, lab.size = 1, add_group.lab_in_pie = T)
}
