% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationplot.R
\name{pLevel}
\alias{pLevel}
\title{Convert numeric p-values into significance levels}
\usage{
pLevel(pval)
}
\arguments{
\item{pval}{pval numeric vector; p-value(s) to be converted}
}
\value{
character vector;
}
\description{
Convert numeric p-values into significance levels
}
