% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{venn_plot}
\alias{venn_plot}
\title{VennPlot for sets no more than 4}
\usage{
venn_plot(
  set_list,
  pal = (ggsci::pal_lancet(palette = c("lanonc"), alpha = 0.6))(9),
  set_name_size = 6,
  text_size = 6,
  show_percentage = F,
  title,
  title_size = 20
)
}
\arguments{
\item{set_list}{a list of vectors with names for each set!}

\item{pal}{color for each set default ggsci::pal_lancet}

\item{set_name_size}{text size for set names, defautl 6}

\item{text_size}{text size for counts of each insects, defautl 6}

\item{show_percentage}{if show the percent of each insects , defautl F}
}
\value{
a ggplot object of venn plot
}
\description{
VennPlot for sets no more than 4
}
\examples{
venn_plot(set_list = list,  pal = c("#0099CC", "#FF6666"), set_name_size = 6, text_size = 6, show_percentage = T, title = 'test', title_size = 30)
}
