% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin.R
\name{violinplot}
\alias{violinplot}
\title{violinplot with data presented as long, do not analysis and provide mean or sd or counts. Suitable for y data type of continuous!}
\usage{
violinplot(
  data,
  x_value,
  y_value,
  group,
  add.box = T,
  box.color = "white",
  box.width = 0.2,
  xlab_level,
  group_level,
  add.p = F,
  method = "t.test",
  title = NULL,
  x_lab = NULL,
  y_lab = "",
  legend.title = "",
  base.size = 18,
  legend.position = "top",
  legend.size = 12,
  xlab.angle = 45,
  pal = (ggsci::pal_lancet(palette = c("lanonc"), alpha = 0.6))(9),
  alpha.violin = 0.8,
  chinese = F,
  plot.theme
)
}
\arguments{
\item{data}{dataframe with long form (raw data, do not analysis!); y continuous data}

\item{x_value}{x column name , data type must be discrete}

\item{y_value}{Y column name,  data type must be continuous for mean/sd}

\item{group}{group column name, data type must be discrete (for filling sub type of the x value)}

\item{add.box}{if adding boxplot in the violinplot, default T}

\item{box.color}{the color of the boxplot ,default white}

\item{box.width}{the width of the boxplot ,default 0.2}

\item{xlab_level}{levels of x lab}

\item{group_level}{levels of group (legend)}

\item{add.p}{if adding p value on the plot, default F, using ggpubr::stat_compare_means}

\item{method}{default t.test}

\item{title}{plot title}

\item{x_lab}{plot x lab}

\item{y_lab}{plot y lab}

\item{legend.title}{plot legend title}

\item{base.size}{base text size of the theme default 18}

\item{legend.position}{legend position default top}

\item{legend.size}{legend text size default 12}

\item{xlab.angle}{if rotate the x labs defualt 45}

\item{pal}{color for filling  default pal_lancet}

\item{alpha.violin}{alpha of the violin, default 0.8}

\item{chinese}{if text in Chinese ,default False}

\item{plot.theme}{plot.theme}
}
\value{
ggplot plot object
}
\description{
violinplot with data presented as long, do not analysis and provide mean or sd or counts. Suitable for y data type of continuous!
}
\examples{
violinplot(data = df, x_value = 'celltype', y_value = 'NES', group = 'group',
                     y_lab = 'NES', legend.title = 'TMB', add.box = F, xlab.angle = 45, pal = c("#D2ADA8", "#8CA3C3"), add.p = T)

}
