% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arranged_barplot.R
\name{waterfallplot}
\alias{waterfallplot}
\title{A special bar plot with ordered bars based on y value, and labs lay on the bars}
\usage{
waterfallplot(
  data,
  x_value,
  y_value,
  fill_group,
  labs,
  fill_level,
  descend = F,
  lab.size = 3,
  pal = (ggsci::pal_lancet(palette = c("lanonc"), alpha = 0.6))(9),
  title = NULL,
  x_lab = NULL,
  y_lab = "",
  legend.title = NULL,
  base.size = 18,
  legend.position = "top",
  legend.size = 12,
  chinese = F
)
}
\arguments{
\item{data}{a dataframe with  x (usually samples , not shown on the plot); y value (numeric); group for filling colors, labs text shown on the bar (not necessary)}

\item{x_value}{column name of x value in the dataframe (sample ids, not shown)}

\item{y_value}{column name of y value in the dataframe}

\item{fill_group}{column name of filling groups}

\item{labs}{column name of lab text shown on the bar ,could be y value or defined as you wish}

\item{fill_level}{levels of fill lab}

\item{descend}{if using desceding order of y}

\item{lab.size}{lab text size default 3}

\item{pal}{color for filling  default pal_lancet}

\item{title}{plot title default NULL}

\item{x_lab}{x axis lab default NULL}

\item{y_lab}{y axis lab default NULL}

\item{legend.title}{legend title default NULL}

\item{base.size}{base text size of the theme default 18}

\item{legend.position}{legend position default top}

\item{legend.size}{legend text size default 12}

\item{chinese}{if text in Chinese ,default False}
}
\value{
a ggplot barplot with order y values ; labs outside the bar(2 sides based on y value, like gsea plot)
}
\description{
A special bar plot with ordered bars based on y value, and labs lay on the bars
}
\examples{
waterfallplot(data = data[1:20,], x_value = 'patientID', y_value = 'TIDE', fill_group = 'group', labs = 'TIDE')
}
