% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationplot.R
\name{cor_matrix}
\alias{cor_matrix}
\title{show correlation plot for multiple variables (more than two vars, show by matrix, not scatter points)}
\usage{
cor_matrix(
  cor.matrix,
  method = "spearman",
  show = "mix",
  shape = "ellipse",
  title = "",
  lab.color = "black",
  lab.size = 1,
  pal = colorRampPalette(c("#2254aa", "white", "#f3261f"))(100)
)
}
\arguments{
\item{cor.matrix}{data frame with different vars: row samples, columns different var names (all in cor analysis)}

\item{method}{cor analyse method can be "pearson", "spearman", "kendall", default spearman}

\item{show}{can only be mix or upper, default mix:with shapes (upper) and cor values (lower)}

\item{shape}{can be 'circle', 'square', 'ellipse', 'number', 'pie', 'shade' and 'color'; default ellipse}

\item{title}{plot title}

\item{lab.color}{color of labs}

\item{lab.size}{text size of labs}

\item{pal}{color bar for cor value}
}
\value{
a cor plot for multi factors
}
\description{
show correlation plot for multiple variables (more than two vars, show by matrix, not scatter points)
}
\examples{
cor_matrix1(cor.matrix = mtcars, show = 'upper', shape = 'pie')
}
